fix(dto) : DTO내 id 값의 자료형 수정

+++ src/main/java/jeje/work/aeatbe/dto/ProductAllergyDTO.java
@@ -7,6 +7,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductAllergyDTO {
-    private int id;
-    private int productId;
+    private Long id;
+    private Long productId;
 }
